Warnungen beseitigt

+++ src/markermodules/bscansegmentation/bscanseglocalop.h
@@ -19,7 +19,7 @@ public:
 	BScanSegLocalOp(BScanSegmentation& parent) : segmentation(parent) {}
 	virtual ~BScanSegLocalOp() {}
 
-	virtual void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, int factor) const = 0;
+	virtual void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, double factor) const = 0;
 	virtual bool drawMarker() const                                 { return true; }
 
 	virtual bool endOnCoord(int x, int y)   = 0;
@@ -63,7 +63,7 @@ public:
 	BScanSegLocalOpPaint(BScanSegmentation& parent) : BScanSegLocalOp(parent) {}
 
 
-	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, int factor) const override;
+	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, double factor) const override;
 	bool drawMarker() const                 override                { return localPaintData.paintMethod != BScanSegmentationMarker::PaintData::PaintMethod::Pen; }
 
 	bool endOnCoord(int /*x*/, int /*y*/)   override                { return false; }
@@ -97,7 +97,7 @@ public:
 	BScanSegLocalOpThreshold(BScanSegmentation& parent) : BScanSegLocalOp(parent) {}
 
 
-	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, int factor) const override;
+	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, double factor) const override;
 
 	bool endOnCoord(int x, int y)           override                { return applyThreshold(x, y); }
 	bool drawOnCoord(int x, int y)          override                { if(applyOnMouseMove) return applyThreshold(x, y); return false; }
@@ -126,7 +126,7 @@ public:
 	BScanSegLocalOpOperation(BScanSegmentation& parent) : BScanSegLocalOp(parent) {}
 
 
-	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, int factor) const override;
+	void drawMarkerPaint(QPainter& painter, const QPoint& centerDrawPoint, double factor) const override;
 
 	bool endOnCoord(int x, int y)            override;
 	bool drawOnCoord(int /*x*/, int /*y*/)   override               { return false; }
